Move navigation tests
